bug 598507 - wrap test harness process execution with 'arch -arch i386' on OS X 10.5. r=catlee, a=blocking
UltraBlame original commit: 0c0e27537b72e2c98f6a65da0599997479d0a786

+++ build/automationutils.py
@@ -36,16 +36,20 @@
 
 
 
-import glob, logging, os, subprocess, sys
+import glob, logging, os, platform, shutil, subprocess, sys
 import re
+from urlparse import urlparse
 
 __all__ = [
   "addCommonOptions",
   "checkForCrashes",
   "dumpLeakLog",
+  "isURL",
   "processLeakLog",
   "getDebuggerInfo",
   "DEBUGGER_INFO",
+  "replaceBackSlashes",
+  "wrapCommand",
   ]
 
 
@@ -68,6 +72,10 @@ DEBUGGER_INFO = {
 
 log = logging.getLogger()
 
+def isURL(thing):
+  """Return True if |thing| looks like a URL."""
+  return urlparse(thing).scheme != ''
+
 def addCommonOptions(parser, defaults={}):
   parser.add_option("--xre-path",
                     action = "store", type = "string", dest = "xrePath",
@@ -79,7 +87,7 @@ def addCommonOptions(parser, defaults={}):
   parser.add_option("--symbols-path",
                     action = "store", type = "string", dest = "symbolsPath",
                     default = defaults['SYMBOLS_PATH'],
-                    help = "absolute path to directory containing breakpad symbols")
+                    help = "absolute path to directory containing breakpad symbols, or the URL of a zip file containing symbols")
   parser.add_option("--debugger",
                     action = "store", dest = "debugger",
                     help = "use the given debugger to launch the application")
@@ -94,6 +102,7 @@ def addCommonOptions(parser, defaults={}):
 
 def checkForCrashes(dumpDir, symbolsPath, testName=None):
   stackwalkPath = os.environ.get('MINIDUMP_STACKWALK', None)
+  stackwalkCGI = os.environ.get('MINIDUMP_STACKWALK_CGI', None)
   
   if testName is None:
     try:
@@ -110,15 +119,41 @@ def checkForCrashes(dumpDir, symbolsPath, testName=None):
       
       subprocess.call([stackwalkPath, d, symbolsPath], stderr=nullfd)
       nullfd.close()
+    elif stackwalkCGI and symbolsPath and isURL(symbolsPath):
+      f = None
+      try:
+        f = open(d, "rb")
+        sys.path.append(os.path.join(os.path.dirname(__file__), "poster.zip"))
+        from poster.encode import multipart_encode
+        from poster.streaminghttp import register_openers
+        import urllib2
+        register_openers()
+        datagen, headers = multipart_encode({"minidump": f,
+                                             "symbols": symbolsPath})
+        request = urllib2.Request(stackwalkCGI, datagen, headers)
+        print urllib2.urlopen(request).read()
+      finally:
+        if f:
+          f.close()
     else:
       if not symbolsPath:
         print "No symbols path given, can't process dump."
-      if not stackwalkPath:
-        print "MINIDUMP_STACKWALK not set, can't process dump."
+      if not stackwalkPath and not stackwalkCGI:
+        print "Neither MINIDUMP_STACKWALK nor MINIDUMP_STACKWALK_CGI is set, can't process dump."
       else:
-        if not os.path.exists(stackwalkPath):
+        if stackwalkPath and not os.path.exists(stackwalkPath):
           print "MINIDUMP_STACKWALK binary not found: %s" % stackwalkPath
-    os.remove(d)
+        elif stackwalkCGI and not isURL(stackwalkCGI):
+          print "MINIDUMP_STACKWALK_CGI is not a URL: %s" % stackwalkCGI
+        elif symbolsPath and not isURL(symbolsPath):
+          print "symbolsPath is not a URL: %s" % symbolsPath
+    dumpSavePath = os.environ.get('MINIDUMP_SAVE_PATH', None)
+    if dumpSavePath:
+      shutil.move(d, dumpSavePath)
+      print "Saved dump as %s" % os.path.join(dumpSavePath,
+                                              os.path.basename(d))
+    else:
+      os.remove(d)
     extra = os.path.splitext(d)[0] + ".extra"
     if os.path.exists(extra):
       os.remove(extra)
@@ -134,7 +169,7 @@ def searchPath(directory, path):
   "Go one step beyond getFullPath and try the various folders in PATH"
   
   newpath = getFullPath(directory, path)
-  if os.path.exists(newpath):
+  if os.path.isfile(newpath):
     return newpath
 
   
@@ -142,7 +177,7 @@ def searchPath(directory, path):
   if not os.path.dirname(path):
     for dir in os.environ['PATH'].split(os.pathsep):
       newpath = os.path.join(dir, path)
-      if os.path.exists(newpath):
+      if os.path.isfile(newpath):
         return newpath
   return None
 
@@ -229,6 +264,7 @@ def processSingleLeakFile(leakLogFileName, PID, processType, leakThreshold):
   seenTotal = False
   crashedOnPurpose = False
   prefix = "TEST-PASS"
+  numObjects = 0
   for line in leaks:
     if line.find("purposefully crash") > -1:
       crashedOnPurpose = True
@@ -270,6 +306,10 @@ def processSingleLeakFile(leakLogFileName, PID, processType, leakThreshold):
         else:
           instance = "instance"
           rest = ""
+        numObjects += 1
+        if numObjects > 5:
+          
+          prefix = "TEST-INFO"
         log.info("%(prefix)s %(process)s| automationutils.processLeakLog() | leaked %(numLeaked)d %(instance)s of %(name)s "
                  "with size %(size)s bytes%(rest)s" %
                  { "prefix": prefix,
@@ -317,3 +357,19 @@ def processLeakLog(leakLogFile, leakThreshold = 0):
         processType = m.group(1)
         processPID = m.group(2)
       processSingleLeakFile(thisFile, processPID, processType, leakThreshold)
+
+def replaceBackSlashes(input):
+  return input.replace('\\', '/')
+
+def wrapCommand(cmd):
+  """
+  If running on OS X 10.5 or older, wrap |cmd| so that it will
+  be executed as an i386 binary, in case it's a 32-bit/64-bit universal
+  binary.
+  """
+  if platform.system() == "Darwin" and \
+     hasattr(platform, 'mac_ver') and \
+     platform.mac_ver()[0][:4] < '10.6':
+    return ["arch", "-arch", "i386"] + cmd
+  
+  return cmd
